feat: Add city field to authFormSchema
The authFormSchema in utils.ts has been updated to include a city field. This change allows for capturing the city information in the authentication form. The city field is optional for the 'sign-in' type and has a maximum length of 50 characters. This addition enhances the completeness and flexibility of the authentication form.