feat: Update Home component to fetch logged-in user asynchronously
The code changes in the Home component update the logic for fetching the logged-in user. Instead of using a hardcoded object, the component now calls the `getLoggedInUser` function asynchronously to retrieve the user data. This change improves the accuracy of the displayed user information and enhances the overall user experience.