feat: Update Copy component to enable copying of titles
This commit updates the Copy component to enable users to copy the titles of items. It adds a copyToClipboard function that utilizes the navigator.clipboard API to write the title to the clipboard. The hasCopied state is used to track whether the title has been copied successfully. This enhancement improves the user experience by providing a convenient way to copy titles for sharing or reference purposes.